new model display method

+++ app/src/main/java/com/example/waitr/HelperShape.kt
@@ -4,8 +4,8 @@ data class HelperShape(
     val id: String = "",
     var height: Int = 0,
     var width: Int = 0,
-    var xPosition: Int = 0,
-    var yPosition: Int = 0
+    var xPosition: Float = 0f,
+    var yPosition: Float = 0f
 ){
     fun toMap(): Map<String, Any?> {
         return mapOf(
